Reorganize metrics.
