+++ lsif/src/server/backend/database.ts
@@ -3,7 +3,7 @@ import { Connection } from 'typeorm'
 import { ConnectionCache, DocumentCache, EncodedJsonCacheValue, ResultChunkCache } from './cache'
 import { hashKey, mustGet } from '../../util'
 import { instrument } from '../../shared/metrics'
-import { databaseQueryDurationHistogram, databaseQueryErrorsCounter } from '../../database.metrics'
+import { databaseQueryDurationHistogram, databaseQueryErrorsCounter } from '../metrics'
 import { DefaultMap } from '../../default-map'
 import { gunzipJSON } from '../../encoding'
 import { isEqual, uniqWith } from 'lodash'
