+++ lsif/src/server/server.ts
@@ -24,7 +24,7 @@ import { Logger } from 'winston'
 import { LsifDump } from '../shared/models/xrepo'
 import { metricsMiddleware } from './middleware/metrics'
 import { pipeline as _pipeline } from 'stream'
-import { queueSizeGauge } from '../server.metrics'
+import { queueSizeGauge } from './metrics'
 import { waitForConfiguration } from '../shared/config/config'
 import { XrepoDatabase } from '../shared/xrepo/xrepo'
 import {
@@ -35,11 +35,7 @@ import {
     CLEAN_FAILED_JOBS_INTERVAL,
     HTTP_PORT,
 } from './settings'
-import {
-    connectionCacheCapacityGauge,
-    documentCacheCapacityGauge,
-    resultChunkCacheCapacityGauge,
-} from './backend/cache.metrics'
+import { connectionCacheCapacityGauge, documentCacheCapacityGauge, resultChunkCacheCapacityGauge } from './metrics'
 
 /**
  * Runs the HTTP server which accepts LSIF dump uploads and responds to LSIF requests.
