+++ lsif/src/worker/importer/importer.ts
@@ -1,7 +1,7 @@
 import { assertId, hashKey, mustGet, readEnvInt } from '../../util'
 import { Correlator, ResultSetData, ResultSetId } from './correlator'
 import { createSqliteConnection } from '../../shared/database/sqlite'
-import { databaseInsertionDurationHistogram, databaseInsertionErrorsCounter } from './importer.metrics'
+import { databaseInsertionDurationHistogram, databaseInsertionErrorsCounter } from '../metrics'
 import { DefaultMap } from '../../default-map'
 import { Edge, MonikerKind, RangeId, Vertex } from 'lsif-protocol'
 import { EntityManager } from 'typeorm'
