+++ lsif/src/worker/worker.ts
@@ -13,7 +13,7 @@ import { ensureDirectory } from '../util'
 import { followsFrom, FORMAT_TEXT_MAP, Span, Tracer } from 'opentracing'
 import { instrument } from '../shared/metrics'
 import { Job } from 'bull'
-import { jobDurationErrorsCounter, jobDurationHistogram } from '../worker.metrics'
+import { jobDurationErrorsCounter, jobDurationHistogram } from './metrics'
 import { Logger } from 'winston'
 import { REDIS_ENDPOINT, STORAGE_ROOT } from './settings'
 import { startMetricsServer } from './server'
